# backend/donald/rest/openapi.yaml
...
openapi: 3.1.0
info:
    title: Donald
    version: 1.0.0
servers:
    - url: http://localhost:8080/v1
paths:
...
    /agent-nodes:
        get:
            tags:
                - AgentNode
            summary: List agent-nodes
            operationId: listAgentNode
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of agent-nodes
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentNodeList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - AgentNode
            summary: Create AgentNode
            operationId: createAgentNode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentNode'
                required: true
            responses:
                "201":
                    description: Created AgentNode
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentNode'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /agent-runs:
        get:
            tags:
                - AgentRun
            summary: List agent-runs
            operationId: listAgentRun
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of agent-runs
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentRunList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - AgentRun
            summary: Create AgentRun
            operationId: createAgentRun
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentRun'
                required: true
            responses:
                "201":
                    description: Created AgentRun
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentRun'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
components:
    schemas:
...
        AgentNode:
            properties:
                agent_label:
                    nullable: true
                    type: string
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                description:
                    nullable: true
                    type: string
                error_message:
                    nullable: true
                    type: string
                finished_at:
                    format: date-time
                    nullable: true
                    type: string
                input_summary:
                    nullable: true
                    type: string
                name:
                    type: string
                node_key:
                    type: string
                node_type:
                    $ref: '#/components/schemas/AgentNodeType'
                    nullable: true
                output_summary:
                    nullable: true
                    type: string
                plan_order:
                    format: int64
                    nullable: true
                    type: integer
                planned:
                    type: boolean
                progress_percent:
                    format: int64
                    nullable: true
                    type: integer
                run_uuid:
                    format: uuid
                    type: string
                started_at:
                    format: date-time
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/AgentNodeStatus'
                status_message:
                    nullable: true
                    type: string
                tool_name:
                    nullable: true
                    type: string
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - run_uuid
                - node_key
                - name
                - status
                - planned
                - created_by
                - updated_by
                - created_at
                - updated_at
            type: object
        AgentNodeList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/AgentNode'
                    type: array
                next_page_token:
                    type: string
            type: object
        AgentNodeStatus:
            enum:
                - invalid
                - not_started
                - in_progress
                - succeeded
                - blocked_on_user_decision
                - blocked_on_missing_data
                - blocked_on_provider_outage
                - failed
                - cancelled
                - skipped
            type: string
        AgentNodeType:
            enum:
                - invalid
                - plan_step
                - tool_call
                - reasoning
                - decision
                - user_interaction
                - subagent_call
                - external_call
                - other
            type: string
        AgentRun:
            properties:
                agent_identifier:
                    nullable: true
                    type: string
                agent_summary:
                    nullable: true
                    type: string
                client_uuid:
                    format: uuid
                    type: string
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                display_summary:
                    nullable: true
                    type: string
                finished_at:
                    format: date-time
                    nullable: true
                    type: string
                graph_revision:
                    format: int64
                    type: integer
                initial_prompt:
                    nullable: true
                    type: string
                last_event_sequence:
                    format: int64
                    type: integer
                last_heartbeat_at:
                    format: date-time
                    nullable: true
                    type: string
                name:
                    nullable: true
                    type: string
                protocol_version:
                    nullable: true
                    type: string
                run_key:
                    type: string
                started_at:
                    format: date-time
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/AgentRunStatus'
                status_message:
                    nullable: true
                    type: string
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - client_uuid
                - run_key
                - status
                - graph_revision
                - last_event_sequence
                - created_by
                - updated_by
                - created_at
                - updated_at
            type: object
        AgentRunList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/AgentRun'
                    type: array
                next_page_token:
                    type: string
            type: object
        AgentRunStatus:
            enum:
                - invalid
                - in_progress
                - failed
                - blocked_on_user_decision
                - blocked_on_missing_data
                - blocked_on_provider_outage
                - succeeded
                - pending
                - cancelled
            type: string
...
        Problem:
            properties:
                detail:
                    type: string
                instance:
                    type: string
                status:
                    type: integer
                title:
                    type: string
                type:
                    type: string
            type: object