# backend/donald/rest/server/list_agent_node.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"log"
	"net/http"

	"go.einride.tech/aip/filtering"
	"go.einride.tech/aip/ordering"

	agent_nodemodule "github.com/nextwave/donald/core/module/agent_node"
	"github.com/nextwave/donald/core/module/agent_node/types"
	main_entity "github.com/nextwave/donald/entity/agent_node"
)

func (h *handlers) ListAgentNode(w http.ResponseWriter, r *http.Request) {
	defaultPageSize := int32(10)
	maxPageSize := int32(100)

	lp, err := parseListParams(r, defaultPageSize, maxPageSize)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	offset, err := parsePageTokenOffset(lp.pageToken)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid page token: %w", err))
		return
	}

	declarations := agent_nodeDeclarations()
	filter, err := filtering.ParseFilter(lp, declarations)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid filter expression: %w", err))
		return
	}

	orderBy, err := ordering.ParseOrderBy(lp)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid order_by expression: %w", err))
		return
	}

	req := types.ListRequest{
		Offset:                offset,
		PageSize:              lp.GetPageSize(),
		Filter:                filter,
		FilteringDeclarations: declarations,
		EnumValues:            agent_nodeEnumValues(),
		OrderBy:               orderBy,
		IncludeFields:         lp.includeFields,
		ExcludeFields:         lp.excludeFields,
	}

	var result types.ListResponse
	if lp.skipCache {
		result, err = h.core.AgentNode().List(r.Context(), req, agent_nodemodule.WithSkipCache())
	} else {
		result, err = h.core.AgentNode().List(r.Context(), req)
	}
	if err != nil {
		// A filter or order_by the clause builder cannot express is the caller's
		// error, not the server's: statusForError turns it into a 400.
		writeProblem(w, statusForError(err), err)
		return
	}

	resp := ListEnvelope[main_entity.AgentNode]{
		Items: result.AgentNode,
	}
	if result.HasNextPage {
		resp.NextPageToken = encodePageToken(offset + int64(lp.GetPageSize()))
	}

	writeJSON(w, http.StatusOK, resp)
}

func agent_nodeDeclarations() *filtering.Declarations {
	declarations, err := filtering.NewDeclarations(
		filtering.DeclareStandardFunctions(),
		filtering.DeclareIdent("true", filtering.TypeBool),
		filtering.DeclareIdent("false", filtering.TypeBool),

		// agent_node

		filtering.DeclareIdent("uuid", filtering.TypeString),

		filtering.DeclareIdent("run_uuid", filtering.TypeString),

		filtering.DeclareIdent("node_key", filtering.TypeString),

		filtering.DeclareIdent("name", filtering.TypeString),

		filtering.DeclareIdent("description", filtering.TypeString),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in agent_nodeEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("node_type", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_TYPE_INVALID", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_TYPE_PLAN_STEP", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_TYPE_TOOL_CALL", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_TYPE_REASONING", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_TYPE_DECISION", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_TYPE_USER_INTERACTION", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_TYPE_SUBAGENT_CALL", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_TYPE_EXTERNAL_CALL", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_TYPE_OTHER", filtering.TypeString),

		filtering.DeclareIdent("agent_label", filtering.TypeString),

		filtering.DeclareIdent("tool_name", filtering.TypeString),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in agent_nodeEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("status", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_STATUS_INVALID", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_STATUS_NOT_STARTED", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_STATUS_IN_PROGRESS", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_STATUS_SUCCEEDED", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_STATUS_BLOCKED_ON_USER_DECISION", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_STATUS_BLOCKED_ON_MISSING_DATA", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_STATUS_BLOCKED_ON_PROVIDER_OUTAGE", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_STATUS_FAILED", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_STATUS_CANCELLED", filtering.TypeString),
		filtering.DeclareIdent("AGENT_NODE_STATUS_SKIPPED", filtering.TypeString),

		filtering.DeclareIdent("status_message", filtering.TypeString),

		filtering.DeclareIdent("error_message", filtering.TypeString),

		filtering.DeclareIdent("progress_percent", filtering.TypeInt),

		filtering.DeclareIdent("plan_order", filtering.TypeInt),

		filtering.DeclareIdent("planned", filtering.TypeBool),

		filtering.DeclareIdent("input_summary", filtering.TypeString),

		filtering.DeclareIdent("output_summary", filtering.TypeString),

		filtering.DeclareIdent("started_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("finished_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("created_by", filtering.TypeString),

		filtering.DeclareIdent("updated_by", filtering.TypeString),

		filtering.DeclareIdent("created_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("updated_at", filtering.TypeTimestamp),
	)
	if err != nil {
		log.Printf("error creating declarations for agent_node: %v", err)
	}
	return declarations
}

// agent_nodeEnumValues maps each enum filter field to every accepted
// spelling of its values and the number the column stores. REST generates no
// protobuf enum types (an app built with --api rest has no proto package at
// all), so this table — not the filter declarations — is what makes an enum
// filter resolvable here.
func agent_nodeEnumValues() map[string]map[string]int64 {
	return map[string]map[string]int64{

		"node_type": {
			"invalid":                          0,
			"AGENT_NODE_TYPE_INVALID":          0,
			"plan_step":                        1,
			"AGENT_NODE_TYPE_PLAN_STEP":        1,
			"tool_call":                        2,
			"AGENT_NODE_TYPE_TOOL_CALL":        2,
			"reasoning":                        3,
			"AGENT_NODE_TYPE_REASONING":        3,
			"decision":                         4,
			"AGENT_NODE_TYPE_DECISION":         4,
			"user_interaction":                 5,
			"AGENT_NODE_TYPE_USER_INTERACTION": 5,
			"subagent_call":                    6,
			"AGENT_NODE_TYPE_SUBAGENT_CALL":    6,
			"external_call":                    7,
			"AGENT_NODE_TYPE_EXTERNAL_CALL":    7,
			"other":                            8,
			"AGENT_NODE_TYPE_OTHER":            8,
		},

		"status": {
			"invalid":                                      0,
			"AGENT_NODE_STATUS_INVALID":                    0,
			"not_started":                                  1,
			"AGENT_NODE_STATUS_NOT_STARTED":                1,
			"in_progress":                                  2,
			"AGENT_NODE_STATUS_IN_PROGRESS":                2,
			"succeeded":                                    3,
			"AGENT_NODE_STATUS_SUCCEEDED":                  3,
			"blocked_on_user_decision":                     4,
			"AGENT_NODE_STATUS_BLOCKED_ON_USER_DECISION":   4,
			"blocked_on_missing_data":                      5,
			"AGENT_NODE_STATUS_BLOCKED_ON_MISSING_DATA":    5,
			"blocked_on_provider_outage":                   6,
			"AGENT_NODE_STATUS_BLOCKED_ON_PROVIDER_OUTAGE": 6,
			"failed":                      7,
			"AGENT_NODE_STATUS_FAILED":    7,
			"cancelled":                   8,
			"AGENT_NODE_STATUS_CANCELLED": 8,
			"skipped":                     9,
			"AGENT_NODE_STATUS_SKIPPED":   9,
		},
	}
}