# backend/donald/rest/server/list_agent_run.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"log"
	"net/http"

	"go.einride.tech/aip/filtering"
	"go.einride.tech/aip/ordering"

	agent_runmodule "github.com/nextwave/donald/core/module/agent_run"
	"github.com/nextwave/donald/core/module/agent_run/types"
	main_entity "github.com/nextwave/donald/entity/agent_run"
)

func (h *handlers) ListAgentRun(w http.ResponseWriter, r *http.Request) {
	defaultPageSize := int32(10)
	maxPageSize := int32(100)

	lp, err := parseListParams(r, defaultPageSize, maxPageSize)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	offset, err := parsePageTokenOffset(lp.pageToken)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid page token: %w", err))
		return
	}

	declarations := agent_runDeclarations()
	filter, err := filtering.ParseFilter(lp, declarations)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid filter expression: %w", err))
		return
	}

	orderBy, err := ordering.ParseOrderBy(lp)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid order_by expression: %w", err))
		return
	}

	req := types.ListRequest{
		Offset:                offset,
		PageSize:              lp.GetPageSize(),
		Filter:                filter,
		FilteringDeclarations: declarations,
		EnumValues:            agent_runEnumValues(),
		OrderBy:               orderBy,
		IncludeFields:         lp.includeFields,
		ExcludeFields:         lp.excludeFields,
	}

	var result types.ListResponse
	if lp.skipCache {
		result, err = h.core.AgentRun().List(r.Context(), req, agent_runmodule.WithSkipCache())
	} else {
		result, err = h.core.AgentRun().List(r.Context(), req)
	}
	if err != nil {
		// A filter or order_by the clause builder cannot express is the caller's
		// error, not the server's: statusForError turns it into a 400.
		writeProblem(w, statusForError(err), err)
		return
	}

	resp := ListEnvelope[main_entity.AgentRun]{
		Items: result.AgentRun,
	}
	if result.HasNextPage {
		resp.NextPageToken = encodePageToken(offset + int64(lp.GetPageSize()))
	}

	writeJSON(w, http.StatusOK, resp)
}

func agent_runDeclarations() *filtering.Declarations {
	declarations, err := filtering.NewDeclarations(
		filtering.DeclareStandardFunctions(),
		filtering.DeclareIdent("true", filtering.TypeBool),
		filtering.DeclareIdent("false", filtering.TypeBool),

		// agent_run

		filtering.DeclareIdent("uuid", filtering.TypeString),

		filtering.DeclareIdent("client_uuid", filtering.TypeString),

		filtering.DeclareIdent("run_key", filtering.TypeString),

		filtering.DeclareIdent("name", filtering.TypeString),

		filtering.DeclareIdent("display_summary", filtering.TypeString),

		filtering.DeclareIdent("agent_summary", filtering.TypeString),

		filtering.DeclareIdent("agent_identifier", filtering.TypeString),

		filtering.DeclareIdent("initial_prompt", filtering.TypeString),

		filtering.DeclareIdent("protocol_version", filtering.TypeString),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in agent_runEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("status", filtering.TypeString),
		filtering.DeclareIdent("AGENT_RUN_STATUS_INVALID", filtering.TypeString),
		filtering.DeclareIdent("AGENT_RUN_STATUS_IN_PROGRESS", filtering.TypeString),
		filtering.DeclareIdent("AGENT_RUN_STATUS_FAILED", filtering.TypeString),
		filtering.DeclareIdent("AGENT_RUN_STATUS_BLOCKED_ON_USER_DECISION", filtering.TypeString),
		filtering.DeclareIdent("AGENT_RUN_STATUS_BLOCKED_ON_MISSING_DATA", filtering.TypeString),
		filtering.DeclareIdent("AGENT_RUN_STATUS_BLOCKED_ON_PROVIDER_OUTAGE", filtering.TypeString),
		filtering.DeclareIdent("AGENT_RUN_STATUS_SUCCEEDED", filtering.TypeString),
		filtering.DeclareIdent("AGENT_RUN_STATUS_PENDING", filtering.TypeString),
		filtering.DeclareIdent("AGENT_RUN_STATUS_CANCELLED", filtering.TypeString),

		filtering.DeclareIdent("status_message", filtering.TypeString),

		filtering.DeclareIdent("graph_revision", filtering.TypeInt),

		filtering.DeclareIdent("last_event_sequence", filtering.TypeInt),

		filtering.DeclareIdent("started_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("finished_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("last_heartbeat_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("created_by", filtering.TypeString),

		filtering.DeclareIdent("updated_by", filtering.TypeString),

		filtering.DeclareIdent("created_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("updated_at", filtering.TypeTimestamp),
	)
	if err != nil {
		log.Printf("error creating declarations for agent_run: %v", err)
	}
	return declarations
}

// agent_runEnumValues maps each enum filter field to every accepted
// spelling of its values and the number the column stores. REST generates no
// protobuf enum types (an app built with --api rest has no proto package at
// all), so this table — not the filter declarations — is what makes an enum
// filter resolvable here.
func agent_runEnumValues() map[string]map[string]int64 {
	return map[string]map[string]int64{

		"status": {
			"invalid":                                     0,
			"AGENT_RUN_STATUS_INVALID":                    0,
			"in_progress":                                 1,
			"AGENT_RUN_STATUS_IN_PROGRESS":                1,
			"failed":                                      2,
			"AGENT_RUN_STATUS_FAILED":                     2,
			"blocked_on_user_decision":                    3,
			"AGENT_RUN_STATUS_BLOCKED_ON_USER_DECISION":   3,
			"blocked_on_missing_data":                     4,
			"AGENT_RUN_STATUS_BLOCKED_ON_MISSING_DATA":    4,
			"blocked_on_provider_outage":                  5,
			"AGENT_RUN_STATUS_BLOCKED_ON_PROVIDER_OUTAGE": 5,
			"succeeded":                                   6,
			"AGENT_RUN_STATUS_SUCCEEDED":                  6,
			"pending":                                     7,
			"AGENT_RUN_STATUS_PENDING":                    7,
			"cancelled":                                   8,
			"AGENT_RUN_STATUS_CANCELLED":                  8,
		},
	}
}